refactor: рефакторинг переменных, типов и утилс // refactoring of variables, types and utils
