+++ src/algorithms/linked-list-with-snapshots/linked-list-with-snapshots.ts
@@ -1,4 +1,16 @@
-import {LinkedListNode} from "../../utils/linked-list-node";
+import {nanoid} from "nanoid";
+
+export class LinkedListNode<T> {
+  letter: T
+  next: LinkedListNode<T> | null;
+  id: string
+
+  constructor(value: T, next?: LinkedListNode<T> | null) {
+    this.letter = value;
+    this.next = (next === undefined ? null : next);
+    this.id = nanoid(5)
+  }
+}
 
 export type TNewSnapList<T> = {
   head: LinkedListNode<T> | null;
