+++ src/components/step-by-step-display/step-by-step-display.tsx
@@ -1,40 +1,49 @@
 import {Queue} from "../../utils/queue";
 import React, {useState} from "react";
 import {DELAY_IN_MS} from "../../constants/delays";
-import {TSnapshot} from "../../types/element-and-snapshot";
 import {useLocation} from "react-router-dom";
-/*import {test} from "../../utils/confetti/confetti";*/
+import {VisualContentList} from "../visual-content-list/visual-content-list";
+import {VisualContentQueue} from "../visual-content-queue/visual-content-queue";
+import {VisualContentString} from "../visual-content-string/visual-content-string";
+import {VisualContentFibonacci} from "../visual-content-fibonacci/visual-content-fibonacci";
+import {VisualContentSorting} from "../visual-content-sorting/visual-content-sorting";
+import {startConfetti} from "../confetti/confetti";
+import {VisualContentStack} from "../visual-content-stack/visual-content-stack";
+import {Buttons} from "../../types/buttons";
+import {Path} from "../../types/path";
 
 type TSteps<T> = {
-  steps: Array<Array<T>>,
+  steps: Array<T>,
   delay?: number,
-  setLoader: React.Dispatch<React.SetStateAction<boolean>>,
-  childComponent: (elementsList: Array<T> ) => JSX.Element
-}
+  setLoader: React.Dispatch<React.SetStateAction<null | Buttons>>,
+};
 
+export const StepByStepDisplay = <T, >({setLoader, steps, delay = DELAY_IN_MS}: TSteps<T>) => {
 
-export const StepByStepDisplay = <T,>({setLoader, steps, delay = DELAY_IN_MS, childComponent}: TSteps<T>) => {
+  let location = useLocation();
 
-  const [arr, setArr] = useState<TSnapshot<T> | null >(null);
-
-  let location = useLocation()
-  console.log(location)
+  const [step, setStep] = useState<null | T>(null);
 
   const stepsQueue = React.useMemo(() => {
     if (steps) {
-      return new Queue<Array<T>>(0, steps)
+      return new Queue<T>(0, steps);
     }
-    return null
-  }, [steps])
+    return null;
+  }, [steps]);
+
+  function extractQueueStep(stepsQueue: Queue<T>) {
+    const stepData = stepsQueue.peak();
+    if (stepData) {
+      const stepCopy = {...stepData};
+      setStep(stepCopy);
+    }
+    stepsQueue.dequeue();
+  }
+
 
   React.useEffect(() => {
     if (stepsQueue) {
-      const stepData:T | T[] | null = stepsQueue.peak()
-      if (Array.isArray(stepData)){
-        const newArr2 = [...stepData];
-        setArr(newArr2)
-      }
-      stepsQueue.dequeue()
+      extractQueueStep(stepsQueue)
     }
   }, [stepsQueue]);
 
@@ -43,31 +52,34 @@ export const StepByStepDisplay = <T,>({setLoader, steps, delay = DELAY_IN_MS, ch
 
     let stepsTimeoutId: NodeJS.Timeout;
 
-    if (arr && stepsQueue && !stepsQueue.isEmpty()) {
-
+    if (stepsQueue && !stepsQueue.isEmpty()) {
       stepsTimeoutId = setTimeout(() => {
-        const stepData = stepsQueue.peak()
-        if (Array.isArray(stepData)){
-          const newArr2 = [...stepData];
-          setArr(newArr2)
-
-        }
-        stepsQueue.dequeue()
+        extractQueueStep(stepsQueue);
       }, delay);
     }
 
-    if (arr && stepsQueue && stepsQueue.isEmpty()) {
-      setLoader(false)
+    if (stepsQueue && stepsQueue.isEmpty()) {
+      setLoader(null);
+      if (location.pathname === Path.string
+        || location.pathname === Path.sorting
+        || location.pathname === Path.fibonacci) {
+        startConfetti();
+      }
     }
 
     return () => {
       clearTimeout(stepsTimeoutId);
     };
-  }, [arr]);
+  }, [step]);
 
   return (
     <>
-      {arr && childComponent(arr)}
+      {step && location.pathname === Path.list && <VisualContentList<T> content={step}/>}
+      {step && location.pathname === Path.queue && <VisualContentQueue<T> content={step}/>}
+      {step && location.pathname === Path.string && <VisualContentString<T> content={step}/>}
+      {step && location.pathname === Path.fibonacci && <VisualContentFibonacci<T> content={step}/>}
+      {step && location.pathname === Path.sorting && <VisualContentSorting<T> content={step}/>}
+      {step && location.pathname === Path.stack && <VisualContentStack<T> content={step}/>}
     </>
   );
 };
