+++ src/types/base-element.ts
@@ -5,7 +5,4 @@ export type CircleBaseElement = Required<CircleProps> & {id: string}
 
 export type ColumnBaseElement = Required<ColumnProps> & {id: string}
 
-export type TSnapshot<T> = Array<T>
-
-export type TSnapshotsList<T> = Array<Array<T>>
