+++ src/types/buttons.ts
@@ -1,0 +1,13 @@
+export enum Buttons {
+  addHead = "addHead",
+  addTail = "addTail",
+  deleteHead = "deleteHead",
+  deleteTail = "deleteTail",
+  addByIndex = "addByIndex",
+  deleteByIndex = "deleteByIndex",
+  clear = "clear",
+  reverse = "reverse",
+  fibonacci = "fibonacci",
+  sortDescending = "sortDescending",
+  sortAscending = "sortAscending"
+}
