+++ src/types/path.ts
@@ -1,0 +1,8 @@
+export enum Path {
+  string = "/string",
+  fibonacci = "/fibonacci",
+  sorting = "/sorting",
+  stack = "/stack",
+  queue = "/queue",
+  list = "/list",
+}
