--- src/utils/linked-list-node.ts
@@ -1,13 +1,0 @@
-import {nanoid} from "nanoid";
-
-export class LinkedListNode<T> {
-  letter: T
-  next: LinkedListNode<T> | null;
-  id: string
-
-  constructor(value: T, next?: LinkedListNode<T> | null) {
-    this.letter = value;
-    this.next = (next === undefined ? null : next);
-    this.id = nanoid(5)
-  }
-}
